requirements.txt
redis == 6.1.0
pydantic == 1.9.0
uvicorn[standart] == 0.22.0
fastapi == 0.73.0
python-dotenv==1.1.0
